add api for admin manage lesson

+++ Services/LessonService.js
@@ -1,9 +1,27 @@
+import mongoose from 'mongoose';
 import Lesson from '../Models/LessonModel.js';
 import RegisterCourseService from './RegisterCourseService.js';
 
 const createLesson = async (data) => {
+    let { title, description, videos, order, courseId } = data;
+
+    if(!title || !description || !order || !courseId) {
+        throw new Error('Missing required fields');
+    }
+
+    order = Number.parseInt(order);
+    if(order === NaN) {
+        throw new Error('Invalid order of lesson in course');
+    }
+    
     try {
-        const lesson = new Lesson(data);
+        const lesson = new Lesson({
+            title,
+            description,
+            videos,
+            order,
+            courseId
+        });
         await lesson.save();
         return lesson;
     } catch (error) {
@@ -11,18 +29,13 @@ const createLesson = async (data) => {
     }
 };
 
-const getAllLessons = async () => {
-    try {
-        return await Lesson.find().populate('courseId');
-    } catch (error) {
-        throw new Error('Error fetching lessons: ' + error.message);
-    }
-};
-
 const getLessonById = async (id) => {
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        throw new Error('Invalid lessonId');
+    }
+
     try {
-        const lesson = await Lesson.findById(id).populate('courseId');
-        if (!lesson) throw new Error('Lesson not found');
+        const lesson = await Lesson.findById(id);
         return lesson;
     } catch (error) {
         throw new Error('Error fetching lesson: ' + error.message);
@@ -30,9 +43,26 @@ const getLessonById = async (id) => {
 };
 
 const updateLesson = async (id, data) => {
+    let { title, description, videos, order, courseId } = data;
+
+    if(!title || !description || !order || !courseId) {
+        throw new Error('Missing required fields');
+    }
+
+    order = Number.parseInt(order);
+    if(order === NaN) {
+        throw new Error('Invalid order of lesson in course');
+    }
+
     try {
-        const lesson = await Lesson.findByIdAndUpdate(id, data, { new: true });
-        if (!lesson) throw new Error('Lesson not found');
+        const lesson = await Lesson.findByIdAndUpdate(id, {
+            title,
+            description,
+            videos,
+            order,
+            courseId
+        }, { new: true });
+
         return lesson;
     } catch (error) {
         throw new Error('Error updating lesson: ' + error.message);
@@ -40,23 +70,32 @@ const updateLesson = async (id, data) => {
 };
 
 const deleteLesson = async (id) => {
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        throw new Error('Invalid lessonId');
+    }
+
     try {
         const lesson = await Lesson.findByIdAndDelete(id);
-        if (!lesson) throw new Error('Lesson not found');
         return lesson;
     } catch (error) {
         throw new Error('Error deleting lesson: ' + error.message);
     }
 };
 
-const getLessonsByCourse = async (userId, courseId) => {
+const getLessonsByCourse = async (userId, role, courseId) => {
     try {
         //Not sign-in
         if(!userId) {
             const lessonWithoutVideos = await Lesson.find({ courseId }).select('-videos.url').sort({ order: 'asc' });
             return lessonWithoutVideos;
         }
-        
+
+        //Admin
+        if(role === 'Admin') {
+            const lessonWithVideos = await Lesson.find({ courseId }).sort({ order: 'asc'});
+            return lessonWithVideos;
+        }
+
         //Sign-in
         const registeredCourse = await RegisterCourseService.getRegisteredCourse(userId, courseId);
         if(registeredCourse && registeredCourse?.status == 'Confirmed') {
@@ -75,7 +114,6 @@ const getLessonsByCourse = async (userId, courseId) => {
 
 export default {
     createLesson,
-    getAllLessons,
     getLessonById,
     updateLesson,
     deleteLesson,
